+++ cypress/integration/collections_spec.js
@@ -630,6 +630,22 @@ describe('Dashboard Collections Page', () => {
       });
     });
 
+    it('Should display Granules based on toggling Search By Prefix', () => {
+      const prefix = 'MOD09GQ.A241';
+      const infix = 'A153';
+      cy.visit('/collections/collection/MOD09GQ/006');
+      cy.get('#chk_isprefixsearch').should('be.checked');
+      cy.get('.search').as('search');
+      cy.get('@search').click().type(prefix);
+      cy.get('.table .tbody .tr').should('have.length', 1);
+      cy.get('@search').click().type(infix);
+      cy.get('.table .tbody .tr').should('have.length', 0);
+      cy.get('#chk_isprefixsearch').click().should('not.be.checked');
+      cy.get('.table .tbody .tr').should('have.length.at.least', 1);
+      cy.get('.table .tbody .tr').eq(0).children('.td').eq(2)
+        .contains(infix);
+    });
+
     it('Should dynamically update menu, sidbar and breadcrumb links with latest filter criteria', () => {
       cy.visit('/collections/all');
       cy.wait('@getCollections');